implemented basic spawning of key prefab
in the earlier phase (which i deleted to start from scratch again), the keys were placed in the scene from the start, and was not dynamically instantiated. this still allowed me to track the object's position and rotation while runtime, but it gave me no option to modify the object's posiiton and rotation when the scene was loaded in again, either when the game was restarted or when the scene switch timer triggers a change in scene load. so i changed the method of having the key predefined in the scene, to dynamically spawning the key on scene load. this, along with allowing me to track and save the object's position and rotation, allowed me to now place the object anywhere when the scene loads next, due to some kind of event trigger. i would just have to place the object, now on the saved position and rotation, rather than modifying the existing position of the object, which would reset to the default one at each load, with the first approach