implemented key spawn and pickup and drop with left click onto wife maze
refactored the entire logic of spawning keys and picking them up for the two mazes. now the husband maze has its own logic for spawning and picking up keys, and the wife maze has its own logic for spawning and pickup and drop for its keys. the keys for each maze are now an entity unique to that scene, and the vectors of the two sets of keys can now be read and modified differently, allowing to have save values for the keys in both the scenes separately, allowing me to load the keys in both the scenes id different positions, and not have the the save values for one of the keys in one scene affect a completely different set of keys in another scene