got the key to be picked up (not really) in the husband maze
i tried multiple methods to get this logic implemented (partly the reason why i took so much time after the due date to get the logic implemented). in the end, i realized that i could just do it the way the guy  in the video does it (i got stuck at this point because the guy in te video had different plans after he had the ability to pick up the object (to just use the object rather that be able to redrop it)). but if forgot that this was exactly why i added a different attribute to his last attribute (he added a value to increment when using it while i added an attribute to the key model (so that i could respawn it near the player to immitate a drop of key from inventory)). so i tried multiple methods to try and parent the object to player, instead of just destroying it like he did, simultaneous to adding the object into the inventory, which made errors to pile up, as prefabs could not be modified as they are the assets, and only their instances in the scene could have its position modified until i realized my mistake, and then finally destroying it. this also brought some errors, but it was just from deleting the key, it no longer had the key to save positions of, so that was just a code structure error, which will be fixed in the next commit