implemented the player being able to drop right in front of him.
i had an error of the player dropping the keys somewhere not around the player. this was because the player object that i was using to calculate the drop point was the actual asset and not the instantiated version of it. this meant that the point was updated once at the start and never changed.