attempted to fix one of the suggested bugs - camera bobbing to much
i tried addressing this error. the plan was and is to separate the cameras from the player object (unparent the cameras from the player body). and for this, i need to spawn the cameras dynamically, and add the same code as the player spawning for the camera. but the biggest mistake i made in this plan, is that i set up a plane in the player, and used the location of the plane to spawn the fps camera, and then, based on that spawn the tps camera. but the plane was inside the prefab, and its poisition in the prefab cannot be used to spawn objects in world space. so the new idea is to use the spawned player object and spawn the camera based on the player instead of based on the plane inside the player.