found and partially fixed bug that i stumbled upon
when I was trying to fix the camera bobbing unnecessarily, as pointed out in one of the tests, i stumbled upon another major bug in the key usage. the keys, if picked up with breaks in between (if we pick each after pausing and unpausing), the list system breaks and keys that have not been dropped yet are dropped by the list. this was mainly because of miscommunication between myself and the video i created this off. i eventually the error, which was actually using a method that was created in the video. the creator of the video had cut parts from the video, and might have accidentally cut that part during editing, but i figured it out through a bunch of logs that i wrote throughout my code to figure where the error was originating from. there is still a little more to work on the keys system, before i can proceed with fixing the usability issue pointed down in the testing, by one of my testers