key pick up not working (not to add to inventory, but for using)
here, now that i have separated the fpc and tpc, inside the spawned fpc, i holding the gameObject that holds the camera and not the actual camera. the raycast can only be fired from a camera, and since the object being referenced is not a camera, the raycast is never being fired, and hence, key not being picked up. to fix this, i need to find a way to access the child of this object (the camera), and reference that in the player pickup drop class